+++ package.json
@@ -15,6 +15,7 @@
     "@mui/icons-material": "^5.14.3",
     "@mui/material": "^5.14.4",
     "axios": "^1.5.0",
+    "body-parser": "^1.20.2",
     "date-fns": "^2.30.0",
     "gray-matter": "^4.0.3",
     "next": "latest",
